move more stuff to ids
